fixed float/double issue

+++ src/RcppExports.cpp
@@ -41,33 +41,33 @@ BEGIN_RCPP
 END_RCPP
 }
 // rdist_cpp
-NumericVector rdist_cpp(NumericMatrix A, String metric, float p);
+NumericVector rdist_cpp(NumericMatrix A, String metric, double p);
 RcppExport SEXP _rdist_rdist_cpp(SEXP ASEXP, SEXP metricSEXP, SEXP pSEXP) {
 BEGIN_RCPP
     Rcpp::RObject rcpp_result_gen;
     Rcpp::RNGScope rcpp_rngScope_gen;
     Rcpp::traits::input_parameter< NumericMatrix >::type A(ASEXP);
     Rcpp::traits::input_parameter< String >::type metric(metricSEXP);
-    Rcpp::traits::input_parameter< float >::type p(pSEXP);
+    Rcpp::traits::input_parameter< double >::type p(pSEXP);
     rcpp_result_gen = Rcpp::wrap(rdist_cpp(A, metric, p));
     return rcpp_result_gen;
 END_RCPP
 }
 // pdist_cpp
-NumericMatrix pdist_cpp(NumericMatrix A, String metric, float p);
+NumericMatrix pdist_cpp(NumericMatrix A, String metric, double p);
 RcppExport SEXP _rdist_pdist_cpp(SEXP ASEXP, SEXP metricSEXP, SEXP pSEXP) {
 BEGIN_RCPP
     Rcpp::RObject rcpp_result_gen;
     Rcpp::RNGScope rcpp_rngScope_gen;
     Rcpp::traits::input_parameter< NumericMatrix >::type A(ASEXP);
     Rcpp::traits::input_parameter< String >::type metric(metricSEXP);
-    Rcpp::traits::input_parameter< float >::type p(pSEXP);
+    Rcpp::traits::input_parameter< double >::type p(pSEXP);
     rcpp_result_gen = Rcpp::wrap(pdist_cpp(A, metric, p));
     return rcpp_result_gen;
 END_RCPP
 }
 // cdist_cpp
-NumericMatrix cdist_cpp(NumericMatrix A, NumericMatrix B, String metric, float p);
+NumericMatrix cdist_cpp(NumericMatrix A, NumericMatrix B, String metric, double p);
 RcppExport SEXP _rdist_cdist_cpp(SEXP ASEXP, SEXP BSEXP, SEXP metricSEXP, SEXP pSEXP) {
 BEGIN_RCPP
     Rcpp::RObject rcpp_result_gen;
@@ -75,7 +75,7 @@ BEGIN_RCPP
     Rcpp::traits::input_parameter< NumericMatrix >::type A(ASEXP);
     Rcpp::traits::input_parameter< NumericMatrix >::type B(BSEXP);
     Rcpp::traits::input_parameter< String >::type metric(metricSEXP);
-    Rcpp::traits::input_parameter< float >::type p(pSEXP);
+    Rcpp::traits::input_parameter< double >::type p(pSEXP);
     rcpp_result_gen = Rcpp::wrap(cdist_cpp(A, B, metric, p));
     return rcpp_result_gen;
 END_RCPP
